Refactor further for better package layout
