+++ pkg/util/hostname/apiserver/hostname.go
@@ -1,0 +1,26 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2019 Datadog, Inc.
+
+// +build kubeapiserver
+// +build !kubelet
+
+package apiserver
+
+import a "github.com/DataDog/datadog-agent/pkg/util/kubernetes/apiserver"
+
+func HostnameProvider() (string, error) {
+	nodeName, err := a.HostNodeName()
+	if err != nil {
+		return "", err
+	}
+
+	clusterName := clustername.GetClusterName()
+	if clusterName == "" {
+		log.Debugf("Now using plain kubernetes nodename as an alias: no cluster name was set and none could be autodiscovered")
+		return nodeName, nil
+	} else {
+		return (nodeName + "-" + clusterName), nil
+	}
+}
