+++ pkg/util/hostname/kubelet/host_alias.go
@@ -3,15 +3,21 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-// +build !kubelet
+// +build kubelet
 
-package hostinfo
+package kubelet
 
 import (
 	"fmt"
+
+	"github.com/DataDog/datadog-agent/pkg/util/hostname/validate"
 )
 
 // GetHostAlias uses the "kubelet" hostname provider to fetch the kubernetes alias
 func GetHostAlias() (string, error) {
-	return "", fmt.Errorf("Kubernetes support not build: couldn't extract a host alias from the kubelet")
+	name, err := HostnameProvider()
+	if err == nil && validate.ValidHostname(name) == nil {
+		return name, nil
+	}
+	return "", fmt.Errorf("Couldn't extract a host alias from the kubelet: %s", err)
 }
