+++ pkg/util/hostname/kubelet/no_host_alias.go
@@ -3,21 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-// +build kubelet
+// +build !kubelet
 
 package hostinfo
 
-import (
-	"fmt"
-
-	"github.com/DataDog/datadog-agent/pkg/util/hostname/validate"
-)
+import "fmt"
 
 // GetHostAlias uses the "kubelet" hostname provider to fetch the kubernetes alias
 func GetHostAlias() (string, error) {
-	name, err := KubeletHostnameProvider()
-	if err == nil && validate.ValidHostname(name) == nil {
-		return name, nil
-	}
-	return "", fmt.Errorf("Couldn't extract a host alias from the kubelet: %s", err)
+	return "", fmt.Errorf("Kubernetes support not build: couldn't extract a host alias from the kubelet")
 }
