--- pkg/util/kubernetes/hostinfo/clustername.go
@@ -1,21 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-2019 Datadog, Inc.
-
-// +build kubelet
-
-package hostinfo
-
-import (
-	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/clustername"
-)
-
-func GetClusterName() string {
-	clusterName := clustername.GetClusterName()
-	if clusterName != "" {
-		return clusterName
-	}
-	clusterName, _ = getNodeClusterNameLabel()
-	return clusterName
-}
